### 2017-05-08 #114 @richfitz
cyphr
`cyphr` provides a set of user-friendly wrappers around `openssl` and `cyphr` to encrypt and decrypt data.

```
Package: cyphr
Title: High Level Enryption Wrappers
Version: 0.1.0
Author: Rich FitzJohn
Maintainer: Rich FitzJohn <rich.fitzjohn@gmail.com>
Description: Encryption wrappers, using low-level support from sodium and
    openssl.
    getPass,
    openssl,
    sodium
VignetteBuilder: rmarkdown,
    knitr
```

https://github.com/richfitz/cyphr

Data analysts who deal with sensitive data, especially those who need to collaborate on it.

The packages `openssl` and `sodium` provide the lower level support; this package builds on those to avoid exposing users to raw vectors and to try and make it easy to do the right thing.  There is also overlap with `gpg` (but again that's lower level)

If either are up for it, Ironholds or hrbrmstr (or someone else with similar security focus) (not @-copied directly to avoid adding them to a thread).
