### 2017-05-16 #114 @maelle
Thanks for your submission @richfitz! :smile_cat:

### Editor checks:

- [x ] **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
- [ x] **License:** The package has a CRAN or OSI accepted license
- [x ] **Repository:** The repository link resolves correctly

---

#### Editor comments

Nothing now that R CMD check succeeded on my weird Windows machine! :wink:

Since you asked me whether the README was clear, I have two comments:

* "To generate keys, you really should read the underling documentation in the sodium or openssl packages!"
are these documentations as user-friendly as cyphr's?

* Maybe add a link to https://github.com/hrbrmstr/rpwnd and maybe a few sentences to explain why R is not safe in order to make "(you're using R, remember)" more explicit.

---

Reviewers: @stephlocke @hrbrmstr
Due date: 2017-06-07 for @hrbrmstr but 2017-07-01 for @stephlocke because of travelling.
