### 2017-05-16 #114 @maelle
Thanks @stephlocke and @hrbrmstr to agree to review this package!

Here are links to our [review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) and to our [reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md)
