### 2017-05-18 #114 @hrbrmstr
Gracias. Starting this a bit later this afternoon for a change of pace from tracking global cybercrime ;-)
