### 2017-07-24 #114 @hrbrmstr
# `cyphr` Package Review

## Package Naming

`cyphr` conforms with the rOpenSci requirements.

## Function/variable naming & general syntax

`cyphr` uses snake case and has no function conflicts with other popular packages. It does have similar function names to those in other packages (e.g. `openssl`) but that should not be an issue as this is meant to be a more user-friendly and centralized wrapper for other encryption functions. Little benefit (if any) would be gained from a `cyphr_` prefix for them. Furthermore, the documentation for the package encourages use of the `cyphr::` full namespace prefix which should futher obviate any possible confusion or collisions.

## README

I have submitted a PR to Rich with one spelling fix and a move of the "Installation" section to the rOpenSci suggested position in README's to conform with the the guidelines.

The README provides an _excellent_ overview the intent, usage & utility of the package.

## Code of Conduct

`cyphr` conforms with the rOpenSci requirements.

## Documentation

All exported package functions are fully documented (in roxygen2 format) with examples.

I have submitted a PR to Rich with a bare-bones `cyphr` package-level documentation snippet. It likely does not need to be more than this since there is extensive documentation for the package in the vignettes that will end up being a "must read" for any new user(s).

`cyphr` contains two, *thorough* vignettes that will be extremely helpful for users who are not cybersecurity experts. Rich does a great job explainin "what", "why", and "how".

## News

I have submitted a PR to Rich with a basic NEWS file (`NEWS.md`).

## Authorship

`cyphr` conforms with the rOpenSci requirements.

## Testing

`cyphr` passed checks on my local macOS and Ubuntu systems and I have reviewed & confirmed the same on appveyor and travis.

`cyphr` uses `testthat` and has an extensive test harness that fully exercises & covers all the funtionality of the package.

`cyphr` has 100% code coverage reported on codecov. `#nocov` is used (appropriately) five (5) times

`cyphr` passes all `goodpractice` checks.

The words/identifiers flagged with `devtools::spell_check()` are all fine.

## Versioning

`cyphr` uses semantic versioning.

## Continuous integration

`cyphr` uses travis and appveyor but there are only Windows and Linux checks. All relevant badges are present.

## Examples

Most users will likely need to reference and rely on the vignettes which (as stated previously) are extensive & comprehensive. Not all functions have examples, but many of them would require duplicating substantial content from the vignettes which seems counter-productive. To that end, the PR with the package-level documentation provides links to the vignettes to further encourage users to review the before using the package.

## Package dependencies

`cyphr` conforms with the rOpenSci requirements.

## Recommended scaffolding

N/A

## Console messages

`cyphr` conforms with the rOpenSci requirements.

## Miscellaneous

`cyphr` conforms with the rOpenSci recommendations.

## General Review

The use of session keys is nothing short brilliant. Rich has also done a great job doing everything possible to help users avoid shooting themselves in the foot (feet, since multiple?). It's hard to stop R from being insecure while running but there's more than sufficient

The "registration" capability will also make it possible for authors of other packages to provide encrypt/decrypt functionality for their import/export functions. It's a very clever "hack".

`openssl` and `sodium` are great packages but `cyhpr` makes cryptography far more accessible to the average user with the high level wrapper functions. The "dropbox" use-case used throughout the vignettes put the package use into a very common context that users should find more than adequate to get up and running if they have sensitive data.

The documentation also provides overall solid security guidance without getting "preachy". I likely would have been a bit more forceful (i.e. said "NEVER use ssh keys without a passphrase!!") but that's likely not going to change behaviour whereas Rich's thoughtful, plain-English description about the threat vector for such a practice should appeal to the vast majority of users.

NOTE: (The following is just for communication and should not prohibit the rest of the onboarding process) I haven't really tried "breaking into" other folks' RStudio Server sessions yet as I really haven't had a use-case (apart from trying to get R environment variables) to do so. Attempting to grift `cyphr` session keys would seem to be the first, real use-case. I'm not sure how likely it is that someone with sensitive data will be using a shared computing environment like RStudio Server, but if it's possible for anyone to connect to an R session in an RStudio Server environment then one could grab the session key. It might be worth cc'ing Kevin (Ushey) on this part of the discussion to see if it's worth the time to try such an attack. I only bring it up as this package is intended to secure data analyses and if there are still open attack vectors we should inform users who likely aren't thinking about them since they are just trying to Get Stuff Done.

## "Safety" Review

(my initial attempt at a set of checks for malicious or user/system-unsafe code which I shld really get into a checklist format for rOpenSci soon. since this package is going to handle sensitive data i thought it would be good to test-drive said checks on it.)

- No directly compiled code in the package so no hidden malicious code at the binary level
- No obfuscated malicious R code
- No network access calls
- `system2()` calls are not obfuscated and are appropriate
- `.onLoad()` functionality is appropriate and non-harmful to end users

I tried to "hack" `rewrite_register()` and was able to successfully remove `utils::read.csv` from the package environment which enabled me to have my stealth-registered `malicious::read.csv` called instead (with a bare call on the order of `cyphr::decrypt(read.csv("/tmp/file.csv"), key)`). It should likely be strongly suggested to the user that they always namespace prefix (`::`) i/o calls in `encrypt()`/`decrypt()` and/or those functions should pretty much not execute w/o prefixed calls. This is not a show-stopper, but I'm trying to think like an attacker who really wants data. There are likely easier ways to get said data, too, but this is a "security" package and one could steal stuff from many places if they convinced folks to use their malicious package (or subverted a legit pkg) in conjunction with this one.

## Code Review

- Good+thorough use of function argument assertions
- The documentation inside the package (internal docs) is good but there are 3 TODOs. None of them appear to be show-stoppers.
- In light of the one "attack" vector listed above, it might be worth having `rewrite_register()` being verbose output-wise when it's called and possibly have `encrypt` and `decrypt` be verbose about which function call it actually ran in the event folks don't namespace prefix (either that or force namespace prefixing). I would have put the messaging in, but I didn't want to impose my will upon Rich.
