### 2017-07-24 #114 @maelle
Thanks a lot for your review @hrbrmstr ! :grin:  Could you add how many hours you spent doing it (for our data collection)?
