### 2017-07-24 #114 @kevinushey
> I tried to "hack" rewrite_register() and was able to successfully remove utils::read.csv from the package environment which enabled me to have my stealth-registered malicious::read.csv called instead (with a bare call on the order of cyphr::decrypt(read.csv("/tmp/file.csv"), key)). It should likely be strongly suggested to the user that they always namespace prefix (::) i/o calls in encrypt()/decrypt() and/or those functions should pretty much not execute w/o prefixed calls. This is not a show-stopper, but I'm trying to think like an attacker who really wants data. There are likely easier ways to get said data, too, but this is a "security" package and one could steal stuff from many places if they convinced folks to use their malicious package (or subverted a legit pkg) in conjunction with this one.

I might be misunderstanding what you're poking at here, but for what it's worth, anyone with access to your R session (ie: malicious R packages) can easily overwrite any function you have with e.g.

```
ns <- asNamespace("utils")
unlockBinding("read.csv", ns)
assign("read.csv", function(...) stop("pwned"), envir = ns)
lockBinding("read.csv", ns)

ns <- as.environment("package:utils")
unlockBinding("read.csv", ns)
assign("read.csv", function(...) stop("pwned"), envir = ns)
lockBinding("read.csv", ns)

read.csv("hello.txt")
```

I think that `cyphr` then has to operate under the assumption that the `rsession` it's running in has not been compromised, as I don't think there's much that can be done here (but please correct me if there's something I might be missing).

On the RStudio side, we'd definitely be interested because we certainly don't want to enable new attack vectors in addition to what's already available on the R side :)
