### 2017-07-24 #114 @hrbrmstr
![image](https://user-images.githubusercontent.com/509878/28535789-9366906c-7073-11e7-8ad1-2a20f241c0f3.png)

I'm speculating (from that really quick view with wireshark) that I'd need the session cookie to be able to do anything real to the rsession itself. I wish I had more cycles to poke at it w/o bugging you @kevinushey  ;-)

so we shld prbly add some guidance around ensuring proper user permissions on the server, not having certain tools installed on said servers and encourage folks to run RStudio Server over TLS if they are indeed working with sensitive data in a shared server environment. that may seem obvious but most folks are just trying to get stuff done vs think about  attack scenarios.
