### 2017-07-24 #114 @kevinushey
Indeed, RStudio uses a secure cookie (established at login) to ensure that RPC requests from the browser are routed to the right place. (This goes along with other various safeguards on the server side, e.g. using unix domain sockets validated against the running user ID to ensure that you and only you can interact with your `rsession` process).
