### 2017-07-24 #114 @hrbrmstr
Cool. Then my concerns are pretty much moot as any other threat modeling is
based on needing to be the user which is totes not a cyphr or rstudio
issue. #ty @kevinushey!

On Mon, Jul 24, 2017 at 1:43 PM, Kevin Ushey <notifications@github.com>
wrote:

> Indeed, RStudio uses a secure cookie (established at login) to ensure that
> RPC requests from the browser are routed to the right place. (This goes
> along with other various safeguards on the server side, e.g. using unix
> domain sockets validated against the running user ID to ensure that you and
> only you can interact with your rsession process).
>
