### 2017-07-25 #114 @maelle
Great to see such an animated discussion in this thread!

@richfitz for info you can add a badge for rOpenSci peer-review to `cyphr` README via
