### 2017-07-26 #114 @stephlocke
---

- Well documented and consistent package!
- I created a PR for some typos and to get through the CRAN policy of referring to other packages within apostrophes
- @hrbrmstr's comments provide an excellent review of the security aspects of this package
- A potential improvement to the documentation would be some diagrams, however, the vignettes do a great job anyway
