### 2017-07-28 #114 @richfitz
Thanks for the reviews and especially with PRs!  I will probably get through these in detail next week
