### 2017-08-21 #114 @richfitz
Thanks for the prod @maelle - things refuse to settle down at work!

I've merged the two PRs and added the badge to the README.  The other substantive suggestions from the reviewers are:

* _to add some diagrams_ (@stephlocke) - this is a great suggestion, but realistically is unlikely to happen.  My last attempt at this required significant amounts of elaboration in a talk and there's an art to getting this done.  I will create an issue on the repo and tag it "help wanted" in case someone with more graphics skills than me would like to help.  If @stephlocke or the editors feel strongly at this I can have a go, but it's likely to be a bit slow.
* _There are three TODO's_ (@hrbrmstr) - I had totally forgotten about these; one of them I think I should look into fixing (authenticated decryption with openssl).  This should not take too long to implement - hopefully by the end of the week.  The half-pairs one I think no longer makes sense, and the path improving one I also think can be skipped.  I will tidy them out of the code though.
*

> I tried to "hack" rewrite_register() and was able to successfully remove utils::read.csv from the package environment which enabled me to have my stealth-registered malicious::read.csv called instead (with a bare call on the order of cyphr::decrypt(read.csv("/tmp/file.csv"), key)). It should likely be strongly suggested to the user that they always namespace prefix (::) i/o calls in encrypt()/decrypt() and/or those functions should pretty much not execute w/o prefixed calls. This is not a show-stopper, but I'm trying to think like an attacker who really wants data. There are likely easier ways to get said data, too, but this is a "security" package and one could steal stuff from many places if they convinced folks to use their malicious package (or subverted a legit pkg) in conjunction with this one.

I think this falls into the situation where we must rely on all loaded packages not being malicious; there's nothing stopping a package adding a load hook that harvests ssh keys as they are unlocked.  This is consistent with the points that Kevin raises above.  I could move some of the unlocking into C at the cost of maintainability and clarity, but I think it would still be trivially circumvention by a malicious package.

At the same time, I think it would be interesting to explore what could be done to extend the `lockEnvironment()` approach that R has to namespacing to some sort of code signing (hey ho notary) to prevent tampering more generally.  Then with something like that in place, we could replace the general pattern of a package environment for caching/extension with one that does verification.  I'm not sure what bits of the puzzle would be needed, but this would require serious changes to base R I think.  Obviously this falls out of scope of the package, but it seems something worth thinking about
