### 2017-08-21 #114 @maelle
Thanks @richfitz for the update and especially for having carried this out while being busy! I'll wait for you to say you've done that last thing (authenticated decryption with openssl) before asking for the reviewers' approval.

I'm also thinking of how to keep this security discussion going once we'll have closed this issue, maybe via Discuss / a channel in the Slack?
