### 2017-08-30 #114 @richfitz
OK, I have made the last set of changes and pushed to github!  Please let me know if there's anything I have missed
