### 2017-09-30 #114 @maelle
@hrbrmstr could you have a look at @richfitz's answer and tell us whether it addresses all your comments? Thanks!
