### 2017-10-19 #114 @sckott
hey @hrbrmstr - did Rich address all your concerns?  yea or nah?
