### 2017-10-29 #114 @maelle
Not really part of the review, my "dummy-friendly" issues

* https://github.com/richfitz/cyphr/issues/13

* https://github.com/richfitz/cyphr/issues/14
