### 2017-11-12 #114 @hrbrmstr
Official response: The author has responded to my review and made changes to my satisfaction. I recommend approving this package

Commentary: I agree w/Rich that there are definitely items I mentioned that are out of scope for this package (since they cannot be solved by the package alone). I've Evernoted this issue and added it to the overall "security & safety TODO" for R work.
