### 2017-11-13 #114 @maelle
Thanks @hrbrmstr! 😸

@jeroen we won't need your review after all, sorry but thanks for your motivation!

@richfitz your package is now approved, thanks a lot to @stephlocke and @hrbrmstr for their reviews!

To-dos:
```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
```
- [] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)
