### 2017-11-28 #114 @richfitz
Sure - I'll do that now.  This is going into ropensci or ropenscilabs?
