### 2017-11-28 #114 @maelle
Cool! ropensci, we no longer have the transitional onboarding at ropenscilabs. :-)
